==> code/1_preprocess/1_yelp/1_parsing_yelp.ipynb <==
# Preprocessing and Parsing Raw Data

This notebook is concerned with parsing and then preprocessing the raw data from the `data/raw` folder, and then storing the new format (csv's) in the `data/processed` folder.

# Importing the relevant libraries
import pandas as pd
import os
import json
import gzip

 ** Amazon Prime Pantry Data **

data = []
with gzip.open('../../data/raw/amazon_prime_pantry.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))
    
# total length of list, this number equals total number of products
print("Number of reviews:", len(data))

# first row of the list
print("Example of the data (first row):", data[0])

df = pd.DataFrame.from_dict(data)  # Parsing the data as a pandas dataframe 
df.head(5)  # Displayin the first 5 rows of the dataframe below

# Saving the data to the processed folder
df.to_csv('../../data/processed/amazon_prime_pantry.csv')

 ** Amazon Gourmet Food and Groceries Data **

data = []
with gzip.open('../../data/raw/amazon_grocery_gourmet.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))
    
# total length of list, this number equals total number of products
print("Number of reviews:", len(data))

# first row of the list
print("Example of the data (first row):", data[0])

df = pd.DataFrame.from_dict(data)  # Parsing the data as a pandas dataframe 
df.head(5)  # Displaying the first 5 rows of the dataframe below

# Saving the data to the processed folder
df.to_csv('../../data/processed/amazon_grocery_gourmet.csv')

 ** Yelp Data **

yelp_file = open('../../data/raw/yelp_all', 'r')  # Opening the file

lines = yelp_file.readlines()  # Storing each line as an element in a list
print(lines[0])  # printing the first line

# Creating a list of dictionaries (each element is a row)
rows = []
for line in lines:
    current_row_list = line.split("\t")
    current_row_dict = {
        "productID": current_row_list[0],
        "fakeLabel": current_row_list[1],
        "date": current_row_list[2],
        "reviewText": current_row_list[3]
    }
    rows.append(current_row_dict)
print(rows[0])  # Printing the first row

df = pd.DataFrame.from_dict(rows)  # Parsing the data as a pandas dataframe 
df.head(5)  # Displaying the first 5 rows of the dataframe below

# Saving the data to the processed folder
df.to_csv('../../data/processed/yelp_all.csv')
